More granular errors (actually only one more).
Main goal of this is to distinguish whether connection
to the controller (signal channel) failed or if that succeeded,
but could not get a media node for the connection.

In the current scheme of things, if websocket dial fails,
a simple http.Get is done on `/rtc/validate` endpoint to understand the
error better.

Just adding one more error return type.
I think if the error is `ErrCannotConnectSignal`, that is the one case
where the media node could not be found.

Also logging the websocket dial error and HTTP response as that method
can return the HTTP response (if any) along with error `ErrBadHandshake`
for things like redirect, auth etc. Want to observe what that returns
and check if we can do an early return for dial failures.
NOTE: Things could change between the web socket dial and the second
http.Get. So, it woul dbe nice to see if we can make a decision purely
based on websocket dial return value.