Set lastTS to last sample timestamp.
Was using adjusted last based on number of samples.
But, that does not work if somebody is blasting samples.
For example, if somebody is blasting 10 seconds worth of samples in 1
second, the adjusted last TS would have moved past the incoming
timestamp after the first sample and the subsequent samples would have
been dropped as out-of-order.

The change here is to use the incoming time stamp as is for lastTS.

But, this does fail in the scenario of
- Blast 10 seconds worth of audio in one second.
- 20 seconds later, send another sample.
In this case, the gap should be 10 seconds, but this change will make it
look like 19 seconds.

Need to think more about handling that case. Avoiding the out-of-order
strangle for now.