Remove option to not write silence on no data. (#677)
* Remove option to not write silence on no data.

Although a nice option, unfortunately it does not work with media-sdk
writer pipeline. The writer pipeline submits a sample into the track's
packetiser with just the duration of the sample. So, if samples are not
provided periodically, the RTP time stamp calculation and therefore the
RTCP Sender Report time stamp calculation will both be wrong.

We can expose this option when the pipeline can accommodate gaps. Till
then, it is not good to have this in the public API as it would lead to
incorrect use/behaviour.

* clean up

* update README