Use packet receive time to calculate PTS.
The pipeline clock is anchored to when the first track of the
synchronizer group is innitialised. And all packets should have PTS with
reference to that. That clock is set to when the first packet of any
track in the synchronizer group is read off the wire (non-padding
packet). So, (packet_receive_time - pipeline_start_time) should be used
to get the correct PTS.

A burst of packets is an interesting case.
- smooth flow of packets
- 3 seconds of gap
- a burst of three seconds packets

In this case, the following will happen
- old packet drop threshold will kick in and drop some of those packets.
- estimated PTS will be higher by upto (3 seconds -
  old_packet_treshold) for the burst. But, they should still be
  acceptable and pushed with proper PTS, i. e. PTS calculated using RTP
  timestamp diff should be correct.

Have to ensure that different thresholds do not render the pipeline drop
tracks.