Initialize should be called with the packet reception time (#762)
Adding private function that does almost the same thing as the existing initialze, with the diff that it takes what it used to be now from the packet reception time. Main reason is that initialize might be delayed (until stable sequence of packets arrive), if time.Now is used on the first packet which was gated for a while - it will induce an offset.
Using packet reception time might have as a result that some other track initialized the syncrhonizer main start point to a time t1 that comes after a time of reception of this track first packet. This will in turn make base offset negative - and potentially first few PTSs returned by GetPTS.
For now these negative PTS could be ignored at the caller site before pushing a packet to the gst pipeline. Ideally - these are not ignored but a new offset is generated for the given source with a base time that would reflect the negative offset and pushed to the gst appsource.