Don't report drift that's not applied (#768)
The initial idea behind invoking callback regardless of whether the drift was applied or not was logging. Now that we log not acceptable and high drift values - there is no reason to invoke the callback for the drift that's not applied. Makes data cleaner and prevents acting on something that was essentially skipped.