Revive a track that constantly falls behind pipeline running time (#766)
* Revive a track that constantly falls behind pipeline running time

The second attempt to try to address issues with tracks whose calculated PTS continuously end up too far in the past comparing to the the media pipieline time.
Adding an option to set a callback to get a pipeline running time and max allowed latency comparing to it. If for whichever reason a track is falling behind the running time - latency continuoulsy for 10s - the workaround will pull that track back to the present moment.
Perfect sync can't be maintained after the rebasing but at least the track will be preserved.