Add support for AVCC reader for H.264 publish.
Thanks to Sugun Segu@Oxa for the patch.

With AnnexB, when reading from socket, have to wait for H.264 start
code. That adds a frame of delay as previous frame cannot be dispatched
till the next frame's start code is available.

With AVCC, there is a header length prefix which allows reading a frame
and dispatching it without waiting for next frame start. Streaming
cameras support the RTP mode output format which allows this.

Ideally, we could get this mode of operation into pion/webrtc as well,
but that would be longer route. As we want to make a CLI release with
this, picking up the patch in the SDK.