Do not log duplicate sender report drop.
It is possible to see duplicate sender report at track synchronizer
because of how RTCP compound packets are handled. pion/srtp will forward
the full compound RTCP packet to all the read streams. The synchronizer
fields RTCP packets for all SSRCs it is managing, i. e. all the tracks
it is trying to synchronize. So, it will receive the same compound RTCP
packet once per SSRC contained in the compound packet.